Move OptimizelyUserContext and OptimizelyDecision interfaces to shared_types
