+++ packages/optimizely-sdk/lib/optimizely_decision/index.ts
@@ -13,23 +13,7 @@
  * See the License for the specific language governing permissions and      *
  * limitations under the License.                                           *
  ***************************************************************************/
-import OptimizelyUserContext from '../optimizely_user_context';
-
-export interface OptimizelyDecision {
-  variationKey: string | null;
-  // The boolean value indicating if the flag is enabled or not
-  enabled: boolean;
-  // The collection of variables associated with the decision
-  variables: { [variableKey: string]: unknown };
-  // The rule key of the decision
-  ruleKey: string | null;
-  // The flag key for which the decision has been made for
-  flagKey: string;
-  // A copy of the user context for which the decision has been made for
-  userContext: OptimizelyUserContext;
-  // An array of error/info messages describing why the decision has been made.
-  reasons: string[];
-}
+import { OptimizelyUserContext, OptimizelyDecision } from '../shared_types';
 
 export function newErrorDecision(key: string, user: OptimizelyUserContext, reasons: string[]): OptimizelyDecision {
   return {
