+++ packages/optimizely-sdk/lib/shared_types.ts
@@ -249,3 +249,37 @@ export interface OptimizelyConfig {
   revision: string;
   getDatafile(): string;
 }
+
+export interface OptimizelyUserContext {
+  getUserId(): string;
+  getAttributes(): UserAttributes;
+  setAttribute(key: string, value: unknown): void;
+  decide(
+    key: string,
+    options: OptimizelyDecideOption[]
+  ): OptimizelyDecision;
+  decideForKeys(
+    keys: string[],
+    options: OptimizelyDecideOption[],
+  ): { [key: string]: OptimizelyDecision };
+  decideAll(
+    options: OptimizelyDecideOption[],
+  ): { [key: string]: OptimizelyDecision };
+  trackEvent(eventName: string, eventTags?: EventTags): void;
+}
+
+export interface OptimizelyDecision {
+  variationKey: string | null;
+  // The boolean value indicating if the flag is enabled or not
+  enabled: boolean;
+  // The collection of variables associated with the decision
+  variables: { [variableKey: string]: unknown };
+  // The rule key of the decision
+  ruleKey: string | null;
+  // The flag key for which the decision has been made for
+  flagKey: string;
+  // A copy of the user context for which the decision has been made for
+  userContext: OptimizelyUserContext;
+  // An array of error/info messages describing why the decision has been made.
+  reasons: string[];
+}
